REFACT: Added the SWAN access key for the domain into the www/[domain]/config.json file. There is now a single access key for each domain rather than a single access key shared by all domains in the demo.
Consolidated the requests to the SWAN end points into a single set of methods in the common project. Reduces duplication and enables the access key to be added to the SWAN request in a single central method.

Fixed a bug when personalization was turned off and the publisher page would not appear.