FEAT/CLEANUP/OPTIM: Reduced the number of redirects to the smallest possible.
The Return URL for the publisher is added as a query string parameter to the CMP dialogue request. This avoids reliance on the browser implementing the Rererer header correctly.

Added support for the CMP and the publisher to use difference SWAN Operators (access nodes). The demo publisher now passes this parameter along with the return URL to the CMP. The CMP sets these values in the state array of the SWIFT operation. When the update is performed the URL and access node held in the state are used for the final operation that updates the SWAN data and returns to the publisher.

General code tidy up.