CLEANUP: Standardized on the final naming.
Secure Web Identifier (SWID) is now used in place of Common Browser ID or CBID. This is applied across the solution and will require existing cookies to be reset.

Allow and preferences were used interchangeably. Pref is now used for cookies and parameters.

Stopped and stop were used interchangeably. Stop is now used for all cookies and parameters.