FEAT: SWAN Builder and SWAN ID (#18)
* FEAT: Added a connection structure to facilitate access to the SWAN data layer.

Removed the term ImpressionID and settled on swan.ID as the term and structure name. Avoids naming conflicts with OpenRTB.

* OPTIM: Modified the functions that return the data and not a URL to directly return the data from the connection.

* FEAT: Added a new CMP Sirdata. Connected biscuit-news.uk to the Sirdata CMP.

* BUG: CMP dialog model not preserved when reseting SWID

* CLEANUP: Modified the email handler to point to the path /update from the CMP host domain. This avoids combing user interface functionality with the response from the an email link.

The handler (cmp.handlerUpdate) that processes the email link now creates the update operation using the parameters in the query string. The changes to swanop now ensure a new SWID is added to the operation if one is not provided. Therefore if a SWID already exists in the browser it is not overridden. If the browser is yet to be configured then the new SWID will be used.

* CLEANUP: Modified the email templates to pass integer position values.

Used model as the prefix for the email model.

Fixed bug that showed the wrong data in the email.

* BUG: Modified the storage operations to provide URLs as strings to ensure a consistent query string parameter or path is used to append encrypted data to.

* UPDATE: Used the accessor methods of the SWAN connector project.

* REF: Added salt-go repo as a submodule

* REF: Update salt-js submodule reference

Co-authored-by: James Rosewell <james@51degrees.com>